backend/requirements.txt
asgiref==3.3.4
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2
django-braces==1.14.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-oauth-toolkit==1.3.3
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
idna==2.10
jwcrypto==0.8
oauthlib==3.1.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.4
